+++ app/src/main/res/navigation/main_navigation.xml
@@ -517,7 +517,7 @@
             app:argType="com.infomaniak.mail.data.models.Attachment$AttachmentType" />
         <argument
             android:name="intentType"
-            app:argType="com.infomaniak.mail.utils.extensions.AttachmentExt$AttachmentIntentType" />
+            app:argType="com.infomaniak.mail.utils.extensions.AttachmentExtensions$AttachmentIntentType" />
     </dialog>
 
     <dialog
